Do not use case but functions to handle yield result. In this way is possibile to handle setups where notify/build function never returns an error.
On some usage scenarios (for example building up a map from other variables in the yield block) errors cannot simply happen, but the current implementation forces the user to always have the success and error paths in the yield block. Basically, during the macro expansion, the user provided yield block is always case-matched with two conditions, but if the block does not have a valid path to error, dialyzer will warn about that.