Pass a function instead of a code block to EventSource.notify/2 and EventSource.build/2.
Reduce macro size by delegating implementation to a private module.

Macro cannot be removed completely because we still need to get the caller in order to complete the event with source. As a side effect, we have available also the function, file and line so will be possible to expand the event source informations in the future.